1.fixed transitive vulnerable 2.delete modules, such as zeus-lang、zeus-osf and zeus-jvm 3.adapt the version dependencies of Spring Boot and Spring Framework to meet the requirements of the JDK11 environment
