+++ zeus-demo/src/main/java/com/iterlife/zeus/demo/XmlApplication.java
@@ -1,7 +1,7 @@
 package com.iterlife.zeus.demo;
 
-import com.iterlife.spring.boot.demo.bean.User;
-import com.iterlife.spring.boot.demo.factory.UserFactoryBean;
+import com.iterlife.zeus.demo.entity.User2;
+import com.iterlife.zeus.demo.factory.UserFactoryBean;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
@@ -27,8 +27,8 @@ public class XmlApplication {
         log.info(applicationContext.getBean("&userFactoryBean2").toString());
         log.info(((UserFactoryBean) applicationContext.getBean("&userFactoryBean2")).getObject().toString());
 
-        ((User) applicationContext.getBean("user1")).printAnnotation();
-        ((User) ((UserFactoryBean) applicationContext.getBean("&userFactoryBean2")).getObject()).printAnnotation();
+        ((User2) applicationContext.getBean("user1")).printAnnotation();
+        ((User2) ((UserFactoryBean) applicationContext.getBean("&userFactoryBean2")).getObject()).printAnnotation();
     }
 
 }
