+++ zeus-demo/src/main/java/com/iterlife/zeus/demo/entity/User1.java
@@ -14,12 +14,6 @@ public class User1 {
 
     public int age;
 
-    public User1(String id, String name, int age) {
-        this.id = id;
-        this.name = name;
-        this.age = age;
-    }
-
     public String getId() {
         return id;
     }
