+++ zeus-demo/src/main/java/com/iterlife/zeus/feature/JdkDynamicProxyTest.java
@@ -1,4 +1,4 @@
-package com.iterlife.zeus.lang.feature;
+package com.iterlife.zeus.feature;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
