+++ zeus-demo/src/main/java/com/iterlife/zeus/feature/ThreadLocalTest.java
@@ -1,4 +1,4 @@
-package com.iterlife.zeus.lang.feature;
+package com.iterlife.zeus.feature;
 
 import java.util.ArrayList;
 import java.util.List;
