+++ zeus-demo/src/main/java/com/iterlife/zeus/feature/ThreadPoolExecutorTest.java
@@ -1,4 +1,4 @@
-package com.iterlife.zeus.lang.feature;
+package com.iterlife.zeus.feature;
 
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.LinkedBlockingQueue;
