+++ zeus-demo/src/main/java/com/iterlife/zeus/jvm/PassiveReferenceTest.java
