+++ zeus-demo/src/main/java/com/iterlife/zeus/osf/Object1.java
