+++ zeus-demo/src/main/java/com/iterlife/zeus/osf/Object2.java
