--- zeus-lang/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>com.iterlife</groupId>
-        <artifactId>zeus</artifactId>
-        <version>1.0.0-RELEASE</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>zeus-lang</artifactId>
-    <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-    </properties>
-</project>
