+++ zeus-leetcode/pom.xml
@@ -9,7 +9,9 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
+    <name>zeus-leetcode</name>
     <artifactId>zeus-leetcode</artifactId>
+    <packaging>jar</packaging>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
